Function References

Physical chemistry parameters of compounds

NMRHamiltonian.PhysicalParamsType — Type

struct PhysicalParamsType{T}
- H_IDs::Vector{Int}
- H_inds_sys::Vector{Vector{Int}}
- cs_sys::Vector{Vector{T}}
- H_inds_singlets::Vector{Vector{Int}}
- cs_singlets::Vector{T}
- J_inds_sys::Vector{Vector{Tuple{Int,Int}}}
- J_inds_sys_local::Vector{Vector{Tuple{Int,Int}}}
- J_vals_sys::Vector{Vector{T}}
- ME::Vector{Vector{Vector{Int}}}

The chemical shift and J-coupling values with respect to nuclei labels for a molecule. Based on this information, an algorithm is used to create magnetic equivalence information in ME.

Details ≡≡≡≡≡≡≡≡

H_IDs::Vector{Int}

Numerical (positive integer) numbering of nuclei for a molecule entry. This numbering is taken from a file, so it is not guaranteed to be consecutive nor start at 1.

H_inds_sys::Vector{Vector{Int}}

Nested list of nuclei labels in consecutive positive integers, which NMRHamiltonian refers to as nuclei indices. The element at [i] is a list of nuclei indices for the i-th non-singlet spin system.

cs_sys::Vector{Vector{T}}

Nested list of non-singlet chemical shifts. The element at [i] is a list of chemical shift values for the i-th non-singlet spin system.

H_inds_singlets::Vector{Vector{Int}}

Nested list of nuclei labels in consecutive positive integers, which NMRHamiltonian refers to as nuclei indices. The element at [i] is a list of nuclei indices that contribute to the same i-th singlet.

cs_singlets::Vector{T}

List of chemical shift values for singlets.

J_inds_sys::Vector{Vector{Tuple{Int,Int}}}

Nested list of pairs of nuclei indices. The element at [i] is a list of J-coupling indices for the i-th non-singlet spin system. J-coupling index is a pair of nuclei indices.

J_inds_sys_local::Vector{Vector{Tuple{Int,Int}}}

Nested list of pairs of nuclei indices. The element at [i] is a list of J-coupling indices for the i-th non-singlet spin system. J-coupling index is a pair of nuclei indices. The nuclei indices are re-numbered to start from 1 for each spin system, which NMRHamiltonian refers to as local nuclei indices.

Internally, spin Hamiltonian matrices for each spin system use this particular indexing scheme in their construction.

J_vals_sys::Vector{Vector{T}}

Nested list of J-coupling values. The element at [i] is a list J-coupling values for the i-th non-singlet spin system.

ME::Vector{Vector{Vector{Int}}}

The element at [i][k] is a list of nuclei indices that are magnetically equivalent. The indices in this list collectively make up the k-th magnetically equivalent group in the i-th spin system. The nuclei indices are re-numbered to start from 1 for each spin system, which NMRHamiltonian refers to as local nuclei indices.

Note that NMRHamiltonian defines magnetically equivalent groups must have two or more nuclei, and these equivalence are determined only based on the J-coupling and chemical shift values provided in the corresponding file for the molecule entry at hand.

NMRHamiltonian.getphysicalparameters — Function

getphysicalparameters(
    ::Type{T},
    target_entries::Vector{String},
    H_params_path::String,
    molecule_mapping_file_path;
    unique_cs_digits::Int = 6,
) where T <: AbstractFloat

Assemble the J-coupling and chemical shift values from JSON files for the specified molecules.

Nomenclature

One set of J-coupling and one set of chemical shift values for a molecule is what NMRHamiltonian.jl calls a molecule entry.

It is common to see different J-coupling values of the same molecule reported in literature for similar experimental conditions. NMRHamiltonian.jl requires one to record each instance as separate molecule entry.

Every molecule entry is recorded as a separate JSON file with some arbitrary file name the data collector wants to use. An additional name-mapping JSON file is needed to translate what the NMRHamiltonian.jl user wants to label the entries to what is the data collector chose to name the JSON files. The name-mapping JSON file is likely to require manual set up.

Inputs

::Type{T} – The AbstractFloat datatype to use for storing floating-point data. A viable input here is Float64, which specifies double precision floating-point numbers are to be used.

target_entries – list of molecule entry names.

H_params_path – path to the directory that contain the J-coupling and chemical shift JSON files. Each file corresponds to one entry of a molecule, and has a set of chemical shifts and J-coupling information in the JSON dictionary format. ID1 and ID2 are JSON dictionary keys that specify the spin nucleus label for J-coupling values, and ID is the nucleus label for chemical shift values. The following is an example of the JSON format for a L-Histidine entry:

{
       "J-coupling": [
                       {
                            "ID2": 13,
                            "ID1": 12,
                          "value": -15.350518
                       },
                       {
                            "ID2": 16,
                            "ID1": 12,
                          "value": 7.63377
                       },
                       {
                            "ID2": 16,
                            "ID1": 13,
                          "value": 5.029267
                       }
                     ],
   "chemical shift": [
                       {
                          "value": 3.18747,
                             "ID": 12
                       },
                       {
                          "value": 3.26837,
                             "ID": 13
                       },
                       {
                          "value": 7.14078,
                             "ID": 14
                       },
                       {
                          "value": 8.02487,
                             "ID": 15
                       },
                       {
                          "value": 3.9967,
                             "ID": 16
                       }
                     ]
}

molecule_mapping_file_path – the files in H_params_path might not be named in a manner for the user to know which molecule entry corresponds to which JSON file. Therefore, the user should further define a single JSON file that encodes this mapping. Example: if the file bmse000976_simulation_1.json corresponds to the molecule entry "L-Histidine" and the file Histidine_Govindaraju_2000.json corresponds to the molecule entry "L-Histidine - Govindaraju", and suppose the user only wants to target these two entries, i.e. target_entries contain only these two entries. Then the following entry-filename-mapping JSON file should be used, and molecule_mapping_file_path should be the path to this entry-filename-mapping file.

{
    "L-Histidine": {
        "notes": "http://gissmo.bmrb.io/entry/bmse000976/simulation_1",
       "file name": "bmse000976_simulation_1.json"
    },
    "L-Histidine - Govindaraju": {
        "notes": "https://pubmed.ncbi.nlm.nih.gov/10861994/",
       "file name": "Histidine_Govindaraju_2000.json"
    }
}

Optional inputs

unique_cs_digits – Two chemical shift values in the same JSON file are assigned the same chemical shift value to both nuclei if the two values in the file differ less than unique_cs_digits many decimal places.

Outputs

Phys::Vector{PhysicalParamsType{Float64} – list of the data type PhysicalParamsType that contain chemical shift and J-coupling information for each of the molecule entries in target_entries.

dict_molecule_to_filename – the dictionary that maps molecule entries to meta information such as their corresponding JSON filenames

NMRHamiltonian.extractcs — Function

extractcs(Phys::Vector{PhysicalParamsType{T}}) where T

Assemble the J-coupling and chemical shift values from JSON files for the specified molecules.

Construction functions for molecule mixture-related configurations in NMRHamiltonian.jl depend on a provided list of the number of spin systems and singlets for each molecule. The constructor functions can be inferred this information from the output quantities of this function.

Inputs

Phys::Vector{PhysicalParamsType{T}}

Physical parameters for multiple molecule entries.

Outputs

cs_sys_mixture:::Vector{Vector{Vector{T}}}

The set of chemical shift values for non-singlet spin systems for multiple molecule entries. See PhysicalParamsType.

cs_singlets_mixture::Vector{Vector{T}}

Singlet chemical shift values for multiple molecule entries. See PhysicalParamsType.

NMRHamiltonian.simulate — Function

simulate(
    Phys::Vector{PhysicalParamsType{T}},
    molecule_entries::Vector{String},
    fs::T,
    SW::T,
    ν_0ppm::T,
    config::SHConfig;
) where T <: AbstractFloat

Simulate the resonance components' intensities and frequencies, compute order of coherences for each component, and group the components into resonance groups.

Inputs

Phys – A Vector of physical chemistry parameters, e.g. the output of getphysicalparameters().

molecule_entries – A Vector of compound entries.

fs – the sampling frequency in Hz for use in the simulation.

SW – the spectral window in ppm for use in the simulation.

ν_0ppm – the 0 ppm peak frequency in Hz in the spectrum, for use in the simulation.

config – a configuration file of type SHConfig.

Outputs

As::Vector{SHType{Float64}} – the simulated resonance intensities and frequencies, sub-system order of coherences, and resonance groups.

MSPs – the spin Hamiltonian matrices and coherence-related quantities. For diagnostic purposes.

NMRHamiltonian.readbasechemshifts — Function

function readbasechemshifts(P::PhysicalParamsType{T})::Vector{Vector{T}} where T

Outputs shifts, with index structure [spin system][cs index]. cs is in ppm. Singlet spin systems are at the end of shifts.

NMRHamiltonian.writebasechemshifts! — Function

function writebasechemshifts!(P::PhysicalParamsType{T}, shifts::Vector{Vector{T}}) where T

shifts: index structure [spin system][cs index]. cs has to be in ppm. Singlet spin systems are at the end of shifts.

NMRHamiltonian.extractMEnuclei — Function

extractMEnuclei( Phys::Vector{PhysicalParamsType{T}}, ) where T

Outputs: IDs, cs, entry_IDs

IDs::Vector{Vector{Int}}

Each entry contains the set of nuclei labels that is associated with the corresponding entry in cs.

cs::Vector{T}

Each entry contains the chemical shift entry of a magnetically equivalent group of nuclei. The magnetic equivalent is based on the data stored in the ME field of the entries in Phy`.

entry_IDs::Vector{Tuple{Int,Int}}

Each entry contains a pair of integers that is associated with the corresponding entry in cs. The first integer is the molecule entry number. The second is the spin system number that the corresponding nuclei in IDs belongs to.

Simulation

NMRHamiltonian.SHConfig — Type

@kwdef struct SHConfig{T <: Real}
    coherence_tol::T = convert(T, 0.01)
    relative_α_threshold::T = convert(T, 0.005) # initial intensity pruning.
    tol_radius_1D::T = convert(T, 0.1) # strictly between 0 and 1. The lower, the better the approximation, but would a larger partition (i.e. more resonance groups).
    max_deviation_from_mean::T = convert(T, 0.2)
    acceptance_factor::T = convert(T, 0.99)
    total_α_threshold::T = convert(T, 0.01) # final intensity pruning.
end

Configuration settings for the spin Hamiltonian simulation portion of simulate().

The parameters here control the trade-off between the number of coherence difference features, the set of Δc's, and the number of resonance groups, the set of Δc_bar's, that are generated. The more resonance groups, the better the approximation to the ideal model where each Δc is a resonance group. The resultant model becomes more computationally intensive as the number of resonance groups increase.

≡≡≡≡≡≡≡≡ Details ≡≡≡≡≡≡≡≡

coherence_tol::T

This controls the number of Δc features we get, by how far their sum deviate from -1.

relative_α_threshold::T

This controls the number of Δc feature we get, by how intense their corresponding resonance intensity is relative to the maximum intensity within their local spin system.

max_deviation_from_mean::T

This non-negative controls the number of resonance groups we get. The smaller, the more resonance groups, but better approximation to the preliminary model.

acceptance_factor::T

This non-negative controls the number of resonance groups we get. Takes on value strictly between 0 to 1. Tends to create less resonance groups when it is close to 1, but takes longer to compute.

total_α_threshold::T

This non-negative controls the number of resonance groups we get. This dictates the cumulative intensity threshold for resonance groups, as a proportionality factor of the total intensity of the spin system. The least intense groups that cumulatively add to less than this value is discarded, and αs and Ωs and Δc are updated according to the kept components. Takes on value from [0, 1). When set to zero, does not prune any resonance groups of low intensity during the post-processing step.

NMRHamiltonian.SHType — Type

struct SHType{T}

    # resonance components for each spin systems (indexed by first/outer index).
    αs::Vector{Vector{T}}
    Ωs::Vector{Vector{T}}

    Δc::Vector{Vector{Vector{T}}}
    parts::Vector{Vector{Vector{Int}}}
    Δc_bar::Vector{Vector{Vector{T}}}

    # the number of spins for each spin system.
    N_spins_sys::Vector{Int}

    # misc input for the SH simulation.
    fs::T
    SW::T
    ν_0ppm::T
end

Container for the spin Hamiltonian simulation result of a molecule entry. All arrays follow the 1-indexing scheme.

Frequency information about the NMR data and spectrometer can be obtained from these information. Spectrometer frequency (in MHz) is calculated as SW/fs, where SW is in ppm, and fs is in Hz. An example value is 700.14 ≈ 14005.6/20.004.

To convert from Hz to ppm frequency units, and the inverse, use the following anonymous functions:

hz2ppmfunc = uu->(uu - ν_0ppm)*SW/fs
ppm2hzfunc = pp->(ν_0ppm + pp*fs/SW)

Fields ≡≡≡≡≡≡≡≡

αs::Vector{Vector{T}}

Resonance intensities for each spin systems. Index [i][l] refers to the i-th spin system, l-th resonnace component.

Ωs::Vector{Vector{T}}

Resonance frequencies for each spin systems. In units radians. To convert to Hz, divide by 2π. Index [i][l] refers to the i-th spin system, l-th resonnace component.

Δc::Vector{Vector{Vector{T}}}

Index [i][l] is a 1-D array of the Δc feature of the l-th resonance component in the i-th spin system. The Δc feature of a resonance component can be interpreted as the partial contributions to the quantum number of the resonance component from each subsystem of the the composite spin system. Its eleemnts take on values betwee -1 and 1 for spin-1/2 NMR.

parts::Vector{Vector{Vector{Int}}}

For a fixed spin system i, parts[i][k] is a 1-D array of indices that specify which of the resonance components (e.g., which elements of αs[i] and Ωs[i]) belong in the resonance group (i,k). The following creates α_ik, the resonance components for the (i,k)-th resonance group of the molecule entry:

α_ik = αs[parts[i][k]]

The indices in parts follow the 1-indexing scheme.

Δc_bar::Vector{Vector{Vector{T}}}

Index [i][k] is the weighted average of the features of the (i,k)-th resonance group. The features are

Δc_ik = Δc[parts[i][k]]

and the weights are

αs[parts[i][k]]

The length of each Δc_bar[i] in a spin system should be the same, and should be less than or equal to the number of spins in the i-th spin system. It won't be equal if there are exhibits magnetic equivalence between the spins.

N_spins_sys::VectorInt}

The i-th element is the number of spins in the i-th spin system.

αs_singlets::Vector{T}

The i-th element is the resonance intensity for the i-th singlet resonance component.

Ωs_singlets::Vector{T}

The i-th element is the resonance frequency for the i-th singlet resonance component. To convert to Hz, divide by 2π.

fsSampling frequency of the spectrometer, in Hz. This is the rate at which the complex-valued free-induction decay signal is sampled at. An example value is around 14005.6 Hz for a 700 MHz spectrometer.

SWSpectral window in ppm for the spectrometer. An example value is around 20.004 ppm for a 700 MHz spectrometer.

ν_0ppmThe 0 ppm reference frequency, in Hz. An example value is around 10656.01 Hz for a spectrometer with SW ≈ 20.004 ppm and fs ≈ 14005.6 Hz.

Serialization, input/output

NMRHamiltonian.serializemixture — Function

function serializemixture(
    As::Vector{SHType{T}},
) where T

Returns a Dict{Symbol, Any} object, which can then be used with file serialization packages for saving to disk. For use with saveasJSOn(); see its docstring for more details.

NMRHamiltonian.deserializemixture — Function

deserializemixture(W)

W is of a data type that can be addressed via a key (of type Symbol) and returns a value. Example data types that W can be are Dict{Symbol, Any} or a JSON3 Object.

Outputs a Vector{SHType{T}} variable, where T is the data type of the key :fs in W.

NMRHamiltonian.serializephysicalparams — Function

serializephysicalparams(
    Phys::Vector{PhysicalParamsType{T}},
    molecule_entries::Vector{String},
    ) where T

Returns a Dict{Symbol, Any} object, which can then be used with file serialization packages for saving to disk.

NMRHamiltonian.deserializephysicalparams — Function

deserializephysicalparams(W)

W is of a data type that can be addressed via a key (of type Symbol) and returns a value. Example data types that W can be are Dict{Symbol, Any} or a JSON3 Object.

Outputs a Vector{PhysicalParamsType{T}} and a Vector{String} variable. T is the value stored in the :T key in W.

NMRHamiltonian.saveasJSON — Function

saveasJSON( outputfolder::String, filename::String, dict, )

Example setup:

config = HAM.SHConfig{T}(
    coherence_tol = convert(T, 0.01),
    relative_α_threshold = convert(T, 0.005),
    max_deviation_from_mean = convert(T, 0.2),
    acceptance_factor = convert(T, 0.99),
    total_α_threshold = convert(T, 0.01), # final intensity pruning.
)
unique_cs_digits = 6

Phys, As, MSPs = HAM.loadandsimulate(
    T,
    spectrometer_tag,
    molecule_entries,
    H_params_path,
    molecule_mapping_file_path;
    config = config,
    unique_cs_digits = unique_cs_digits,
    
)

Usage with physical parameters:

S_Phys = HAM.serializephysicalparams(Phys, molecule_entries)
HAM.saveasJSON(
    joinpath(save_folder_path, "Phys.json"),
    S_Phys,
)

Usage with simulation results:

S_As = HAM.serializemixture(As)
HAM.saveasJSON(
    joinpath(save_folder_path, "As.json"),
    S_As,
)

NMRHamiltonian.readJSON — Function

readJSON(file_path::String)

Usage:

# save.
S_Phys = HAM.serializephysicalparams(Phys, molecule_entries)
HAM.saveasJSON(
    Phys_filename,
    S_Phys,
)

# load.
dict_Phys = HAM.readJSON(Phys_filename)
Phys_rec, molecule_entries_rec = HAM.deserializephysicalparams(Dict(dict_Phys))