Function References

NMRHamiltonian.getphysicalparameters — Function

getphysicalparameters(
    ::Type{T},
    target_entries::Vector{String},
    H_params_path::String,
    molecule_mapping_file_path;
    unique_cs_atol = 1e-6,
) where T <: AbstractFloat

Assemble the J-coupling and chemical shift values from JSON files for the specified molecules.

Nomenclature

One set of J-coupling and one set of chemical shift values for a molecule is what NMRHamiltonian.jl calls a molecule entry.

It is common to see different J-coupling values of the same molecule reported in literature for similar experimental conditions. NMRHamiltonian.jl requires one to record each instance as separate molecule entry.

Every molecule entry is recorded as a separate JSON file with some arbitrary file name the data collector wants to use. An additional name-mapping JSON file is needed to translate what the NMRHamiltonian.jl user wants to label the entries to what is the data collector chose to name the JSON files. The name-mapping JSON file is likely to require manual set up.

Inputs

::Type{T} – The AbstractFloat datatype to use for storing floating point data. A viable input here is Float64, which specifies double precision floating point numbers are to be used.

target_entries – list of molecule entry names.

H_params_path – path to the directory that contain the J-coupling and chemical shift JSON files. Each file corresponds to one entry of a molecule, and has a set of chemical shifts and J-coupling information in the JSON dictionary format. ID1 and ID2 are JSON dictionary keys that specify the spin nucleus label for J-coupling values, and ID is the nucleus label for chemical shift values. The following is an example of the JSON format for a L-Histidine entry:

{
       "J-coupling": [
                       {
                            "ID2": 13,
                            "ID1": 12,
                          "value": -15.350518
                       },
                       {
                            "ID2": 16,
                            "ID1": 12,
                          "value": 7.63377
                       },
                       {
                            "ID2": 16,
                            "ID1": 13,
                          "value": 5.029267
                       }
                     ],
   "chemical shift": [
                       {
                          "value": 3.18747,
                             "ID": 12
                       },
                       {
                          "value": 3.26837,
                             "ID": 13
                       },
                       {
                          "value": 7.14078,
                             "ID": 14
                       },
                       {
                          "value": 8.02487,
                             "ID": 15
                       },
                       {
                          "value": 3.9967,
                             "ID": 16
                       }
                     ]
}

molecule_mapping_file_path – the files in H_params_path might not be named in a manner for the user to know which molecule entry corresponds to which JSON file. Therefore, the user should further define a single JSON file that encodes this mapping. Example: if the file bmse000976_simulation_1.json corresponds to the molecule entry "L-Histidine" and the file Histidine_Govindaraju_2000.json corresponds to the molecule entry "L-Histidine - Govindaraju", and suppose the user only wants to target these two entries, i.e. target_entries contain only these two entries. Then the following entry-filename-mapping JSON file should be used, and molecule_mapping_file_path should be the path to this entry-filename-mapping file.

{
    "L-Histidine": {
        "notes": "http://gissmo.bmrb.io/entry/bmse000976/simulation_1",
       "file name": "bmse000976_simulation_1.json"
    },
    "L-Histidine - Govindaraju": {
        "notes": "https://pubmed.ncbi.nlm.nih.gov/10861994/",
       "file name": "Histidine_Govindaraju_2000.json"
    }
}

Optional inputs

unique_cs_atol – Two chemical shift values in the same JSON file are assigned the same chemical shift value to both nuclei if the two values in the file are less than unique_cs_atol.

Outputs

Phys::Vector{PhysicalParamsType{Float64} – list of the data type PhysicalParamsType that contain chemical shift and J-coupling information for each of the molecule entries in target_entries.

dict_molecule_to_filename – the dictionary that maps molecule entries to meta information such as their corresponding JSON filenames

NMRHamiltonian.SHConfig — Type

function SHConfig( ::Type{T}; coherencetol = 0.01, relativeαthreshold = 0.01, maxpartitionsizeoffset::Integer = 0, partitionrate::Real = 2.0, #searchθ::Bool = true, #searchγ::Bool = true, #θdefault = 0.0, γbase = 0.1, γrate = 1.05, maxiterγ = 100, fullyconnectedconvexclustering = false, # overides all knn-related optional inputs maxconnectedcomponentsoffset = -1, startingmanualknn = 60, lengthscalebase = 10.0, lengthscalerate = 0.7, mindynamicrange = 0.95, ccgaptol = 1e-8, ccmaxiters = 300, assignmentzerotol = 1e-3, reportγcost::Bool = true, verbose_kernel::Bool = true, )::SHConfig{T} where T

NMRHamiltonian.simulate — Function

simulate(
    part_algs::Vector{Vector{PartitionAlgorithmType}},
    names::Vector{String},
    fs::T,
    SW::T,
    ν_0ppm::T,
    Phys::Vector{PhysicalParamsType{T}},
    configs::Vector{SHsConfigType{T}}
    ) where T <: AbstractFloat

Performs spin hamiltonian simulation and resonance group partitioning of the molecule entries in names.

Inputs

part_algs::Vector{Vector{PartitionAlgorithmType}}

Element [n][i] is the partition algorithm config molecule entry n, non-singlet spin system i.

names::Vector{String}

List of molecule entries.

fsSampling frequency of the spectrometer, in Hz. This is the rate at which the complex-valued free-induction decay signal is sampled at. An example value is around 14005.6 Hz for a 700 MHz spectrometer.

SWSpectral window in ppm for the spectrometer. An example value is around 20.004 ppm for a 700 MHz spectrometer.

ν_0ppmThe 0 ppm reference frequency, in Hz. An example value is around 10656.01 Hz for a spectrometer with SW ≈ 20.004 ppm and fs ≈ 14005.6 Hz.

Phys::Vector{PhysicalParamsType{T}}

List of chemical shift and J-coupling information for each molecule entry.

configs::Vector{SHsConfigType{T}}

List of spin Hamiltonian simulation configurations for each molecule entry.

Outputs

As::SHType{T}

A list of spin Hamiltonian result data structure variables.

Rs::Vector{Vector{PartitionSearchRecord}}

A list of iterative search results for the partitioning algorithm. This variable is only useful if an iterative search was done for regulariation parameter γ or/and kernel hyperparameter θ.

simulate(
    Phys::Vector{PhysicalParamsType{T}},
    molecule_entries::Vector{String},
    fs::T,
    SW::T,
    ν_0ppm::T,
    config::SHConfig;

    θs::Vector{Vector{T}} = defaultspinsysbuffer(convert(T,Inf), Phys), # non-finite value means use search strategy.
    γs::Vector{Vector{T}} = defaultspinsysbuffer(convert(T,Inf), Phys),

    γ_bases = defaultspinsysbuffer(config.γ_base, Phys),
    γ_rates = defaultspinsysbuffer(config.γ_rate, Phys),
    γ_max_iters = defaultspinsysbuffer(config.max_iter_γ, Phys),

    length_scale_bases = defaultspinsysbuffer(config.length_scale_base, Phys),
    length_scale_rates = defaultspinsysbuffer(config.length_scale_rate, Phys),
    min_dynamic_ranges = defaultspinsysbuffer(config.min_dynamic_range, Phys),
) where T <: AbstractFloat

Simulate the resonance components' intensities and frequencies, compute order of coherences for each component, and group the components into resonance groups.

Inputs

Phys – A Vector of physical chemistry parameters, e.g. the output of getphysicalparameters().

molecule_entries – A Vector of compound entries.

fs – the sampling frequency in Hz for use in the simulation.

SW – the spectral window in ppm for use in the simulation.

ν_0ppm – the 0 ppm peak frequency in Hz in the spectrum, for use in the simulation.

config – a configuration file of type SHConfig.

Optional inputs

θs – This is a nested array for manually specifying the weight kernel bandwidth parameter (positive) for the convex clustering weights. The kernel is a square exponential kernel, i.e., exp(-θ*norm(X[i]-X[j])^2) for the edge weight between data points X[i] and X[j]. θs[n][i] is the positive weight kernel bandwidth of your choice for the n-th compound entry, i-th spin system. Use the default (i.e., leave this optional input alone) if you want to use the automated search strategy for the kernel bandwidth.

γs – The sizing of this nested array is the same as θs, but for the sparsity-inducing parameter (positive, smaller means more resonance groups) for the convex clustering optimization problem.

γ_bases – This needs to be assigned to defaultspinsysbuffer(config.γ_base, Phys),, but replace config.γ_base with your choice of the starting γ parameter. For use with the automated search strategy.

γ_rates – This needs to be assigned to defaultspinsysbuffer(config.γ_rate, Phys),, but replace config.γ_base with your choice of the rate of increase for the next candidate γ parameter. For use with the automated search strategy. We've γ_next = γ_bases*γ_rate^(iter-1), where iter is the current search iteration. For use with the automated search strategy.

γ_max_iters – This is similar to θs, but for the maximum allowed iterations in the automated search strategy for γ.

length_scale_bases and length_scale_rates are similar to γ_bases and γ_rates, but for length_scale = 1/(2*θ).

min_dynamic_ranges – The sizing of this nested array is the same as θs, but contain the stopping condition dynamic range value for which the kernel bandwidth automated search strategy terminates upon.

Outputs

As::Vector{SHType{Float64}} – the simulated resonance intensities and frequencies, sub-system order of coherences, and resonance groups.

Rs – the partition tree used to generate the resonance groups. For diagnostic purposes.

MSPs – the spin Hamiltonian matrices and coherence-related quantities. For diagnostic purposes.

NMRHamiltonian.serializemixture — Function

function serializemixture(
    As::Vector{SHType{T}},
) where T

Returns a Dict{Symbol, Any} object, which can then be used with file serialization packages for saving to disk.

NMRHamiltonian.readbasechemshifts — Function

function readbasechemshifts(P::PhysicalParamsType{T})::Vector{Vector{T}} where T

Outputs shifts, with index structure [spin system][cs index]. cs is in ppm. Singlet spin systems are at the end of shifts.

NMRHamiltonian.extractMEnuclei — Function

extractMEnuclei( Phys::Vector{PhysicalParamsType{T}}, ) where T

Outputs: IDs, cs, entry_IDs

IDs::Vector{Vector{Int}}

Each entry contains the set of nuclei labels that is associated with the corresponding entry in cs.

cs::Vector{T}

Each entry contains the chemical shift entry of a magnetically equivalent group of nuclei. The magnetic equivalent is based on the data stored in the ME field of the entries in Phy`.

entry_IDs::Vector{Tuple{Int,Int}}

Each entry contains a pair of integers that is associated with the corresponding entry in cs. The first integer is the molecule entry number. The second is the spin system number that the corresponding nuclei in IDs belongs to.